+++ mteb/models/cohere_models.py
@@ -8,10 +8,7 @@ import torch
 
 from mteb.encoder_interface import PromptType
 from mteb.model_meta import ModelMeta
-from mteb.models.sentence_transformer_wrapper import (
-    get_prompt_name,
-    validate_task_to_prompt_name,
-)
+
 
 from .wrapper import Wrapper
 
@@ -28,7 +25,7 @@ class CohereTextEmbeddingModel(Wrapper):
         self.model_name = model_name
         self.sep = sep
         self.model_prompts = (
-            validate_task_to_prompt_name(model_prompts) if model_prompts else None
+            self.validate_task_to_prompt_name(model_prompts) if model_prompts else None
         )
 
     def _embed(
@@ -60,7 +57,7 @@ class CohereTextEmbeddingModel(Wrapper):
         prompt_type: PromptType | None = None,
         **kwargs: Any,
     ) -> np.ndarray:
-        cohere_task_type = get_prompt_name(self.model_prompts, task_name, prompt_type)
+        cohere_task_type = self.get_prompt_name(self.model_prompts, task_name, prompt_type)
         if cohere_task_type is None:
             # search_document is recommended if unknown (https://cohere.com/blog/introducing-embed-v3)
             cohere_task_type = "search_document"
