+++ mteb/models/nomic_models.py
@@ -10,10 +10,6 @@ from sentence_transformers import SentenceTransformer
 
 from mteb.encoder_interface import PromptType
 from mteb.model_meta import ModelMeta
-from mteb.models.sentence_transformer_wrapper import (
-    get_prompt_name,
-    validate_task_to_prompt_name,
-)
 
 from .wrapper import Wrapper
 
@@ -33,7 +29,7 @@ class NomicWrapper(Wrapper):
         self.model_name = model_name
         self.model = SentenceTransformer(model_name, revision=revision, **kwargs)
         self.model_prompts = (
-            validate_task_to_prompt_name(model_prompts) if model_prompts else None
+            self.validate_task_to_prompt_name(model_prompts) if model_prompts else None
         )
 
     def to(self, device: torch.device) -> None:
@@ -48,7 +44,7 @@ class NomicWrapper(Wrapper):
         batch_size: int = 32,
         **kwargs: Any,
     ):
-        input_type = get_prompt_name(self.model_prompts, task_name, prompt_type)
+        input_type = self.get_prompt_name(self.model_prompts, task_name, prompt_type)
 
         # default to search_document if input_type and prompt_name are not provided
         if input_type is None:
