+++ mteb/models/voyage_models.py
@@ -8,10 +8,7 @@ import numpy as np
 
 from mteb.encoder_interface import PromptType
 from mteb.model_meta import ModelMeta
-from mteb.models.sentence_transformer_wrapper import (
-    get_prompt_name,
-    validate_task_to_prompt_name,
-)
+
 from mteb.requires_package import requires_package
 
 from .wrapper import Wrapper
@@ -86,7 +83,7 @@ class VoyageWrapper(Wrapper):
         self._model_name = model_name
         self._max_tpm = max_tpm
         self.model_prompts = (
-            validate_task_to_prompt_name(model_prompts) if model_prompts else None
+            self.validate_task_to_prompt_name(model_prompts) if model_prompts else None
         )
 
     def encode(
@@ -99,7 +96,7 @@ class VoyageWrapper(Wrapper):
         **kwargs: Any,
     ) -> np.ndarray:
         input_type = (
-            get_prompt_name(self.model_prompts, task_name, prompt_type) or "document"
+            self.get_prompt_name(self.model_prompts, task_name, prompt_type) or "document"
         )
         return self._batched_encode(sentences, batch_size, input_type)
